Moved SnippetsTableModel class to own source file
